mm: rmap: Fix invalid casting from a void to an enumerator.
 fixes:

../mm/rmap.c:1344:25: warning: cast to smaller integer type 'enum ttu_flags' from 'void *' [-Wvoid-pointer-to-enum-cast]
        enum ttu_flags flags = (enum ttu_flags)arg;
                               ^~~~~~~~~~~~~~~~~~~

+++ mm/rmap.c
@@ -1330,7 +1330,7 @@ static int try_to_unmap_one(struct page *page, struct vm_area_struct *vma,
 	unsigned long sh_address;
 	bool pmd_sharing_possible = false;
 	unsigned long spmd_start, spmd_end;
-	enum ttu_flags flags = (enum ttu_flags)arg;
+	enum ttu_flags flags = (enum ttu_flags)(long)arg;
 
 	/* munlock has nothing to gain from examining un-locked vmas */
 	if ((flags & TTU_MUNLOCK) && !(vma->vm_flags & VM_LOCKED))
